backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Auth ───────────────────────────────────────────────────────────────────────
firebase-admin>=6.5.0

# ── Env ────────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── ML — PyTorch (image model) ─────────────────────────────────────────────────
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0

# ── ML — TensorFlow (audio + signature models) ────────────────────────────────
tensorflow>=2.16.0

# ── Audio features ─────────────────────────────────────────────────────────────
librosa>=0.10.0

# ── Image processing ───────────────────────────────────────────────────────────
Pillow>=10.0.0
opencv-python-headless>=4.9.0
numpy>=1.24.0
